Refactor bpm reconciler works
[#165785951](https://www.pivotaltracker.com/story/show/165785951)
