+++ pkg/kube/controllers/boshdeployment/generated_variable_controller.go
@@ -2,7 +2,6 @@ package boshdeployment
 
 import (
 	"context"
-
 	"strings"
 
 	corev1 "k8s.io/api/core/v1"
