+++ pkg/kube/controllers/controllers.go
@@ -31,6 +31,7 @@ import (
 var addToManagerFuncs = []func(context.Context, *config.Config, manager.Manager) error{
 	boshdeployment.AddDeployment,
 	boshdeployment.AddGeneratedVariable,
+	boshdeployment.AddBPM,
 	extendedjob.AddTrigger,
 	extendedjob.AddErrand,
 	extendedjob.AddJob,
